requirements.txt
Flask
sentence-transformers
scikit-learn
numpy
pandas
python-dotenv
google-generativeai
tqdm
